backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.20

# Redis
redis==5.2.1

# Utils
httpx==0.28.1